+++ BankManagementSys/src/main/java/com/example/BankManagementSys/Entities/WithdrawalTransaction.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -15,5 +16,5 @@ import java.util.List;
 public class WithdrawalTransaction extends Transaction{
 
     @Column(name = "withdrawal_amount")
-    private double withdrawalAmount;
+    private BigDecimal withdrawalAmount;
 }
